Thanks 파일 분리 및 위치 적용
